pyproject.toml
# ----------------------------------------
# Build System Configuration
# ----------------------------------------
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# ----------------------------------------
# Project Metadata (PEP 621)
# ----------------------------------------
[project]
name = "analyst_toolkit"
version = "0.1.0"  # Stable release
description = "A comprehensive toolkit for data analysis, validation, and diagnostics."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Garrett Schumacher", email = "your@email.com" }
]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# ----------------------------------------
# Core Dependencies
# ----------------------------------------
dependencies = [
  "pandas",
  "matplotlib",
  "seaborn",
  "ipython",
  "ipywidgets",
  "joblib",
  "PyYAML",
  "thefuzz",
  "XlsxWriter",
  "openpyxl"
]

# ----------------------------------------
# Optional Developer Dependencies
# ----------------------------------------
[project.optional-dependencies]
dev = [
  "pytest",
  "black",
]

# ----------------------------------------
# Setuptools Package Discovery
# ----------------------------------------
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = [
  "analyst_toolkit",
  "analyst_toolkit.*",
  "m00_utils*",
  "m01_diagnostics*",      
  "m02_validation*",       
  "m03_normalization*",    
  "m04_duplicates*",       
  "m05_detect_outliers*",  
  "m06_outlier_handling*", 
  "m07_imputation*",       
  "m08_visuals*",
  "m10_final_audit*"       
]

[project.scripts]
analyst-toolkit = "analyst_toolkit.run_toolkit_pipeline:main"